resolved bugs related to start menu
